listener/requirements.txt
Flask==3.0.0
google-cloud-pubsub==2.21.0
